Fix PHP7 deprecation warning for calling non-static method non-statically
Output looks like:

Deprecated: Non-static method bench::static_method_to_bench() should not
be called statically in /Cache_Lite/Cache/Lite/Function.php on line 140
This is the output of the function static_method_to_bench(23, 66) !
This is the result of the function static_method_to_bench(23, 66) !

Deprecated: Non-static method bench::static_method_to_bench() should not
be called statically in /Cache_Lite/Cache/Lite/Function.php on line 140
This is the output of the function static_method_to_bench(23, 66) !
This is the result of the function static_method_to_bench(23, 66) !

Deprecated: Non-static method bench::static_method_to_bench() should not
be called statically in /Cache_Lite/Cache/Lite/Function.php on line 140
 This is the output of the function static_method_to_bench(23, 66) !
 output : *** foo *** bar *** bar ***
 result : *** foo *** bar *** bar ***'

Given the function name, it seems like it is intended to be static.

Thanks to Remi Collet and Walt Sorenson for noticing the issue.

Signed-off-by: Nishanth Aravamudan <nish.aravamudan@canonical.com>