track_errors configuration is deprecated in PHP7.2.
This message was shown on all of phpt test cases.
Deprecated: Directive 'track_errors' is deprecated in Unknown on line 0

And track_errors configuration is deprecated in PHP 7.2.
refs: http://php.net/manual/en/migration72.deprecated.php#migration72.deprecated.track_errors-and-php_errormsg

But It was used in PHPUnit's PHPUnit_Extensions_PhptTestCase file and
its private variable. So I overwrote by using ini setting.

It is no problem to disable track errors becasuse there was nowhere to
use $php_errormsg variable.
refs: http://php.net/manual/en/errorfunc.configuration.php#ini.track-errors