Add pear-core-minimal dependency, rather than relying on system PEAR
When installing through Composer with no PEAR installation on the
system, the error-handling requires a local copy of 'PEAR.php'.

The other changes allow the unit tests to be run on such a system:
Composer updates the include_path when initialising the autoloader,
so that the include('PEAR.php') works as expected.